requirements.txt
PyOpenGL
WxPython
numpy
pyrr
numpy-stl